+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed the `AssetBundle` property not being set when the texture is loaded from
   an asset bundle.
 * Avoid addon binder errors by properly depending on SharpDX dlls.
+* Fixed a bug where using `TakeTexture` on a texture returned from an asset bundle
+  would result in the asset bundle texture being deleted.
 
 ## v0.0.6
 ### Fixed